pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openshift-ai-observability-summarizer"
version = "0.1.0"
description = "AI Observability Metric Summarizer"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.116.1",
    "pydantic>=2.0.0",
    "pandas>=2.3.1",
    "requests>=2.32.4",
    "scipy>=1.16.0",
    "numpy>=1.24.0",
    "markdown>=3.8.2",
    "dateparser>=1.2.2",
    "llama-stack-client>=0.2.12",
    "fire>=0.7.0",
    "pytest>=8.4.1",
    "fastmcp>=2.11",
    "pytest-cov>=6.2.1",
    "structlog>=25.0.0",
    "trio>=0.30.0",
    "mcp[client]>=1.13.0",
    "anthropic>=0.39.0",
    "openai>=2.6.0",
    "google-generativeai>=0.8.5",
    "mlflow-tracing>=3.1",
]

[dependency-groups]
test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "httpx>=0.24.0",
    "pytest-cov>=4.0.0",
]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "matplotlib>=3.10.3",
    "streamlit>=1.47.0",
    "uvicorn>=0.35.0",
    "weasyprint>=65.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/__init__.py",
    "src/api/report_assets/report_config.py",
    "src/api/report_assets/report_renderer.py",
    "src/ui/ui.py",
]

[tool.coverage.report]

show_missing = false
